fix(em): carry local-search rotation ids and hard assignments in int64
The parent-expanded local layout encodes a child orientation as
`psi_index * n_pixels + healpix_pixel` on the unreduced fine lattice. At
HEALPix order 9 that id space is about 3072 * 3.1e6 = 9.7e9, beyond int32, so
the ids wrapped to negative values, the exact local engine mistook them for
padding and aborted with "selected padded local rotation" on EMPIAR-10202 at
RELION iteration 21 (the first parent-order-8 / child-order-9 pass). Order 8
(1.2e9) was the last order that fit, which is why 10073 (max order 6) and the
order-7-capped 10202 run never hit it.

Keep rotation ids int64 on the host at every boundary: layout builders,
bucket padding, bucket reordering, and the postprocess decode. The RELION
projection cache no longer allocates a dense per-id map (which would also
scale with the id range, tens of GB at order 9); it keeps the sorted unique
ids and buckets translate their ids to compact cache rows on the host with
searchsorted, so the device only ever sees row indices. The JIT signature is
unchanged (the id_map is now an identity over cache rows).

The flattened hard assignment `rotation_id * n_trans + translation` likewise
overflowed int32 (values of about +/-2.1e9 were visible in the saved
per-particle dumps from order 7 on); encode and store it in int64 together
with the best-pose rotation ids. The padded-rotation error now reports the
offending image, slot, and id range.

Tests: child ids above 2**31 at fine order 9 stay positive int64; cache rows
map large ids and padding correctly and reject unknown ids; hard-assignment
encoding is exact in int64. Fast-guard selection + K-class guards + local
crop tests: 83 passed; symmetry, K-class semantics, PPCA/initial-model
layout users: 468 passed.

+++ recovar/em/dense_single_volume/local_layout.py
@@ -285,7 +285,7 @@ def _build_factorized_local_entries(
                     dtype=np.float32,
                 )
 
-            local_ids = (psi_indices[:, None] * n_pixels + dir_indices[None, :]).reshape(-1).astype(np.int32)
+            local_ids = (psi_indices[:, None] * n_pixels + dir_indices[None, :]).reshape(-1).astype(np.int64)
             local_log_prior = (psi_log_prior[:, None] + dir_log_prior[None, :]).reshape(-1).astype(np.float32)
             counts[image_idx] = int(local_ids.shape[0])
             running_offset += int(local_ids.shape[0])
@@ -294,7 +294,7 @@ def _build_factorized_local_entries(
             log_prior_parts.append(local_log_prior)
 
     rotation_ids_flat = (
-        np.concatenate(rotation_ids_parts, axis=0) if rotation_ids_parts else np.zeros(0, dtype=np.int32)
+        np.concatenate(rotation_ids_parts, axis=0).astype(np.int64, copy=False) if rotation_ids_parts else np.zeros(0, dtype=np.int64)
     )
     rotation_log_priors_flat = (
         np.concatenate(log_prior_parts, axis=0) if log_prior_parts else np.zeros(0, dtype=np.float32)
@@ -391,7 +391,7 @@ def _build_parent_expanded_local_entries(
         child_rotations, parent_map, child_ids = oversampled[:3]
         child_mstep_rotations = oversampled[3] if bool(generate_relion_mstep_rotations) else None
         parent_map = np.asarray(parent_map, dtype=np.int64)
-        child_ids = np.asarray(child_ids, dtype=np.int32)
+        child_ids = np.asarray(child_ids, dtype=np.int64)
         child_log_prior = parent_log_prior[parent_map].astype(np.float32, copy=False)
 
         counts[image_idx] = int(child_ids.shape[0])
@@ -413,7 +413,7 @@ def _build_parent_expanded_local_entries(
             posterior_ids_parts.append(parent_ids[parent_map].astype(np.int32, copy=False))
 
     rotation_ids_flat = (
-        np.concatenate(rotation_ids_parts, axis=0) if rotation_ids_parts else np.zeros(0, dtype=np.int32)
+        np.concatenate(rotation_ids_parts, axis=0).astype(np.int64, copy=False) if rotation_ids_parts else np.zeros(0, dtype=np.int64)
     )
     rotation_log_priors_flat = (
         np.concatenate(log_prior_parts, axis=0) if log_prior_parts else np.zeros(0, dtype=np.float32)
@@ -630,7 +630,7 @@ def build_local_hypothesis_layout(
                 per_image=True,
                 grid_metadata=grid_metadata,
             )
-            local_ids = np.asarray(local_ids, dtype=np.int32).reshape(-1)
+            local_ids = np.asarray(local_ids, dtype=np.int64).reshape(-1)
             local_log_prior = np.asarray(local_log_prior[0], dtype=np.float32).reshape(-1)
             counts[image_idx] = int(local_ids.shape[0])
             offsets[image_idx + 1] = offsets[image_idx] + local_ids.shape[0]
@@ -638,7 +638,7 @@ def build_local_hypothesis_layout(
             log_prior_parts.append(local_log_prior)
 
         rotation_ids_flat = (
-            np.concatenate(rotation_ids_parts, axis=0) if rotation_ids_parts else np.zeros(0, dtype=np.int32)
+            np.concatenate(rotation_ids_parts, axis=0).astype(np.int64, copy=False) if rotation_ids_parts else np.zeros(0, dtype=np.int64)
         )
         rotation_log_priors_flat = (
             np.concatenate(log_prior_parts, axis=0) if log_prior_parts else np.zeros(0, dtype=np.float32)
@@ -808,7 +808,7 @@ def build_local_adaptive_pass2_hypothesis_layout(
     for image_idx, sig_samples in enumerate(significant_sample_indices):
         parent_start = int(parent_layout.rotation_offsets[image_idx])
         parent_stop = int(parent_layout.rotation_offsets[image_idx + 1])
-        local_parent_ids = np.asarray(parent_layout.rotation_ids_flat[parent_start:parent_stop], dtype=np.int32)
+        local_parent_ids = np.asarray(parent_layout.rotation_ids_flat[parent_start:parent_stop], dtype=np.int64)
         local_parent_log_prior = np.asarray(
             parent_layout.rotation_log_priors_flat[parent_start:parent_stop],
             dtype=np.float32,
@@ -863,7 +863,7 @@ def build_local_adaptive_pass2_hypothesis_layout(
         oversampled_rots = np.asarray(oversampled_rots, dtype=np.float32)
         oversampled_mstep_rots = np.asarray(oversampled_mstep_rots, dtype=np.float32)
         parent_map = np.asarray(parent_map, dtype=np.int32)
-        oversampled_rot_indices = np.asarray(oversampled_rot_indices, dtype=np.int32)
+        oversampled_rot_indices = np.asarray(oversampled_rot_indices, dtype=np.int64)
         parent_posterior_ids = unique_rot[parent_map].astype(np.int32, copy=False)
 
         if use_full_candidate_mask:
@@ -902,7 +902,7 @@ def build_local_adaptive_pass2_hypothesis_layout(
         else np.zeros((0, 3, 3), dtype=np.float32)
     )
     rotation_ids_flat = (
-        np.concatenate(rotation_ids_parts, axis=0) if rotation_ids_parts else np.zeros(0, dtype=np.int32)
+        np.concatenate(rotation_ids_parts, axis=0).astype(np.int64, copy=False) if rotation_ids_parts else np.zeros(0, dtype=np.int64)
     )
     posterior_ids_flat = (
         np.concatenate(posterior_ids_parts, axis=0) if posterior_ids_parts else np.zeros(0, dtype=np.int32)
@@ -1130,7 +1130,7 @@ def build_pass2_hypothesis_layout(
         )
         oversampled_rots = np.asarray(oversampled_rots, dtype=np.float32)
         parent_map = np.asarray(parent_map, dtype=np.int32)
-        oversampled_rot_indices = np.asarray(oversampled_rot_indices, dtype=np.int32)
+        oversampled_rot_indices = np.asarray(oversampled_rot_indices, dtype=np.int64)
         coarse_parent_ids = unique_rot[parent_map].astype(np.int32, copy=False)
 
         if rotation_log_prior_np is None:
@@ -1173,7 +1173,7 @@ def build_pass2_hypothesis_layout(
         np.concatenate(rotations_parts, axis=0) if rotations_parts else np.zeros((0, 3, 3), dtype=np.float32)
     )
     rotation_ids_flat = (
-        np.concatenate(rotation_ids_parts, axis=0) if rotation_ids_parts else np.zeros(0, dtype=np.int32)
+        np.concatenate(rotation_ids_parts, axis=0).astype(np.int64, copy=False) if rotation_ids_parts else np.zeros(0, dtype=np.int64)
     )
     posterior_ids_flat = (
         np.concatenate(posterior_ids_parts, axis=0) if posterior_ids_parts else np.zeros(0, dtype=np.int32)
@@ -1280,7 +1280,7 @@ def bucket_local_hypothesis_layout(
                 (batch_size, int(bucket_size), 3, 3),
             ).copy()
             padded_mstep_rotations = padded_rotations.copy()
-            padded_rotation_ids = np.full((batch_size, int(bucket_size)), -1, dtype=np.int32)
+            padded_rotation_ids = np.full((batch_size, int(bucket_size)), -1, dtype=np.int64)
             padded_log_prior = np.full((batch_size, int(bucket_size)), -1e30, dtype=np.float32)
             padded_mask = np.zeros((batch_size, int(bucket_size)), dtype=bool)
             padded_posterior_ids = (
